Release 0.0.3

+++ update.sh
@@ -4,12 +4,12 @@ set -e
 
 scriptname=update
 test -n "$tag" || . ./vars.sh "$@"
-#test -n "$api_user" || get_env
+test -n "$api_user" || get_env
 
 type err >/dev/null 2>&1 || { . ./util.sh; }
 
 
-log "Updates"
+info "Updates"
 
 case "$image_type" in
 
@@ -19,7 +19,7 @@ case "$image_type" in
       trueish "$Update_Plugins" && {
 
         # Upgrade plugins once jenkins is online, and restart again
-        log "Updating plugins..."
+        info "Updating plugins..."
         ./jenkins-user-script.sh dotmpe-init wait_for_jenkins
         ./jenkins-user-script.sh dotmpe-init update_plugins
         ./jenkins-user-script.sh dotmpe-init wait_for_jenkins
@@ -28,7 +28,7 @@ case "$image_type" in
       trueish "$Update_Customizations" && {
 
         # Customize header logo+title
-        log "Customizing..."
+        info "Customizing..."
         docker exec -ti $cname \
             /opt/dotmpe/docker-jenkins/init.sh customize
       }
@@ -36,14 +36,14 @@ case "$image_type" in
 
       trueish "$Update_Packages" && {
 
-        log "Adding more packages"
+        info "Adding more packages"
         docker exec -u root -ti $cname npm install -g grunt-cli stylus recess
       }
 
 
       trueish "$Update_Credentials" && {
 
-        log "Updating credentials"
+        info "Updating credentials"
 
         docker cp credentials.json \
           $cname:/opt/dotmpe/docker-jenkins/credentials.json
@@ -56,11 +56,11 @@ case "$image_type" in
               Update_Credentials_Clear=$Update_Credentials_Clear \
               Build_Credentials_Json_File="'/opt/dotmpe/docker-jenkins/credentials.json'" \
           && {
-            log "Credentials succesfully updated"
+            info "Credentials succesfully updated"
             break;
           } || {
             update_cred_retries=$(( $update_cred_retries - 1 ))
-            log "Credentials groovy script failed, retries: $update_cred_retries"
+            info "Credentials groovy script failed, retries: $update_cred_retries"
             sleep 30
           }
         done
@@ -75,19 +75,27 @@ case "$image_type" in
 
         # Cloud configuration
 
-        sed 's/\-dev\>/-'$env'/g' clouds.json > clouds-$env.json
-        docker cp clouds-$env.json \
-          $cname:/opt/dotmpe/docker-jenkins/docker-cloud.json
-        rm clouds-$env.json
+        cld_json=clouds-$env.json
 
-        test -n "$DOCKER_URL" || err "DOCKER_URL missing" 1
+        sed 's/\-dev\>/-'$env'/g' clouds.json \
+	  | sed 's/\<localhost\>/'"$DCKR_HOST"'/g' > $cld_json
+
+        # FIXME: update JSON with arrays
+        #{
+        #  echo
+        #} | jsotk.py update $cld_json -
+
+        docker cp $cld_json $cname:/opt/dotmpe/docker-jenkins/docker-cloud.json
+        rm $cld_json
+
+        test -n "$DCKR_SWARM_URL" || error "DCKR_SWARM_URL missing" 1
 
         env=$env tag=$tag cname=$cname \
         ./update-cli-groovy.sh  configure-docker-cloud  \
             Update_Cloud_Name="'$(hostname) Docker Machine'" \
             Update_Cloud_Clear=$Update_Cloud_Clear \
             Build_Docker_Cloud_Json_File="'/opt/dotmpe/docker-jenkins/docker-cloud.json'" \
-            Swarm_Master_URL="'$DOCKER_URL'"
+            Swarm_Master_URL="'$DCKR_SWARM_URL'"
 
         docker exec $cname rm /opt/dotmpe/docker-jenkins/docker-cloud.json
       }
@@ -97,11 +105,11 @@ case "$image_type" in
 
         . ./script/sh/create-jenkins-nodes.sh
 
-        log "Adding node"
+        note "Adding node"
         test -n "$DOTMPE_SLAVE_PORT" -a -n "$DOTMPE_SLAVE_HOST" || {
-          err "Missing env vars ($DOTMPE_SLAVE_HOST:$DOTMPE_SLAVE_PORT)" 1
+          error "Missing env vars ($DOTMPE_SLAVE_HOST:$DOTMPE_SLAVE_PORT)" 1
         }
-        #log "Adding/updating nodes (slaves.tab)"
+        #info "Adding/updating nodes (slaves.tab)"
         #update_nodes_from_table slaves.tab
 
         nodeid=${hostname}-jenkins-slave-dotmpe-$env
@@ -111,8 +119,8 @@ case "$image_type" in
             HOST="$DOTMPE_SLAVE_HOST" PORT="$DOTMPE_SLAVE_PORT" \
             CREDENTIALS_ID='jenkins-ssh-slave-passwd-credentials' \
             LABEL='ubuntu trusty debian' USER='jenkins'  \
-            && log "Added node $nodeid" \
-            || err "Error adding node $nodeid"
+            && info "Added node $nodeid" \
+            || error "Error adding node $nodeid"
 
         nodeid=${hostname}-jenkins-slave-dotmpe-dind-$env
         generate_node $nodeid \
@@ -121,9 +129,45 @@ case "$image_type" in
             HOST="$DOTMPE_SLAVE_HOST" PORT="$DOTMPE_DIND_SLAVE_PORT" \
             CREDENTIALS_ID='jenkins-ssh-slave-passwd-credentials' \
             LABEL='ubuntu trusty debian dind' USER='jenkins'  \
-            && log "Added node $nodeid" \
-            || err "Error adding node $nodeid"
+            && info "Added node $nodeid" \
+            || error "Error adding node $nodeid"
 
+      }
+
+
+      trueish "$Update_Views" && {
+
+        # Create some practical views
+        note "Updating views..."
+
+        # XXX: see CLI init below, test again
+        #info "Moving fresh cutom/views into container"
+        #docker exec -ti $cname rm -rf $chome/custom/views || printf ""
+        #docker cp custom/views/ $cname:$chome/custom/views
+        #info "Copy OK"
+
+        info "Custom dir at container:"
+        docker exec $cname ls -la $chome/custom/views
+
+        for p in custom/views/*.xml
+        do
+
+          name=$(basename $p .xml)
+          case "$name" in
+            Cat )
+                grep -q categorized-view plugins.txt || continue
+              ;;
+            ls )
+                grep -q extra-columns plugins.txt || continue
+              ;;
+          esac
+
+          ./script/sh/create-jenkins-view.sh $p \
+            && info "Created custom view '$name'" || {
+              error "Error creating custom view '$name' ($p) ($?)"
+            }
+
+        done
       }
 
 
@@ -141,12 +185,12 @@ case "$image_type" in
 
         # Use JTB in container to setup some jobs
         test -n "$Build_Offline" \
-          && docker exec -ti $cname bash -c 'cd $JTB_HOME/ && git checkout dev' \
-          || docker exec -ti $cname bash -c 'cd $JTB_HOME/ && git checkout dev && git pull origin dev'
+          && docker exec -ti $cname bash -c 'cd $JTB_SRC_DIR/ && git checkout dev' \
+          || docker exec -ti $cname bash -c 'cd $JTB_SRC_DIR/ && git checkout dev && git pull origin dev'
 
         . ./script/sh/create-jenkins-jobs.sh
 
-        log "Updating jobs (projects.tab)"
+        note "Updating jobs (projects.tab)"
         generate_jobs_from_tab projects.tab
 
         # TODO cleanup old JJB/JTB scripts
@@ -155,47 +199,9 @@ case "$image_type" in
       }
 
 
-      trueish "$Update_Views" && {
-
-        # Create some practical views
-        log "Updating views..."
-
-        # XXX: see CLI init below, test again
-        #log "Moving fresh cutom/views into container"
-        #docker exec -ti $cname rm -rf $chome/custom/views || printf ""
-        #docker cp custom/views/ $cname:$chome/custom/views
-        #log "Copy OK"
-
-        for p in custom/views/*.xml
-        do
-
-          name=$(basename $p .xml)
-          case "$name" in
-            Cat )
-                grep -q categorized-view plugins.txt || continue
-              ;;
-            ls )
-                grep -q extra-columns plugins.txt || continue
-              ;;
-          esac
-
-          log "Creating custom view '$name'"
-
-          ./script/sh/create-jenkins-view.sh $p || {
-            err "Error creating '$name' ($p)"
-          }
-
-          #docker exec -ti $cname jenkins-cli delete-view $name 2>/dev/null || printf ""
-          #docker exec -ti $cname \
-          #    /opt/dotmpe/docker-jenkins/init.sh init_view $name
-
-        done
-      }
-
-
       trueish "$Build_Triggers" && {
 
-        log "Triggering builds (build-triggers.tab)"
+        info "Triggering builds (build-triggers.tab)"
         . ./script/sh/trigger-jenkins-job.sh
         trigger_builds_from_table build-triggers.tab
